3105-modificación de componentes y añadir ConfirmationService

+++ src/app/features/admin/manage-therapies-page/manage-therapies-page.component.ts
@@ -7,6 +7,7 @@ import { Button } from 'primeng/button';
 import { TherapyService } from '@core/services/therapy.service';
 
 import { Therapy } from '@core/models/therapy.model';
+import { ConfirmationService } from 'primeng/api';
 
 @Component({
   selector: 'manage-therapies-page',
@@ -21,13 +22,24 @@ import { Therapy } from '@core/models/therapy.model';
 })
 export class ManageTherapiesPageComponent {
   private therapyService = inject(TherapyService);
+  private confirmationService = inject(ConfirmationService);
 
   therapies: Therapy[] = this.therapyService.getAll();
 
   deleteTherapy(id: string): void {
-    this.therapyService.delete(id);
+    this.confirmationService.confirm({
+      header: 'Eliminar terapia',
 
-    this.therapies = this.therapyService.getAll();
+      message: '¿Seguro que quieres eliminar esta terapia?',
+
+      icon: 'pi pi-exclamation-triangle',
+
+      accept: () => {
+        this.therapyService.delete(id);
+
+        this.therapies = this.therapyService.getAll();
+      },
+    });
   }
 
   cloneTherapy(id: string): void {
